Update sass to 3.4.x and compass to 1.0.3
This gets Sass on the latest version that Compass supports so that we
have more Sass features and bug fixes. This specifically allows for
nested selectors like:

```
.test
   color: red

   &__other
      color: blue
```

to work.

Because we rely on Compass for right now, I didn't update us to the
latest sass gem ver or switch to sassc yet.

Fixes https://github.com/bigcartel/dugway/issues/184