Preset styles support (#196)
Adds necessary validations for our new Preset styles picker. This change ensures that when presets are specified, all of the data is there, it's in the correct format, and it matches up with the other font/color settings in settings.json

Also makes a change to allow for an empty images folder, since themes can be packaged with 0 images.